[debug] Reduce HelixClusterManagerTest matrix; extract FAR tests
Replaces the maxParallelForks=4 experiment (which underperformed serial
on CI) with a structural matrix reduction.

1. Drop fullAutoCompatible from HelixClusterManagerTest's @Parameterized
   matrix. Rows [9] and [10] of the historical matrix collapsed to [0]
   and [6] when the column was removed — a lossless 4-axis dedupe.
   The matrix shrinks 11 → 9 rows, ~18% fewer constructor invocations
   across all ~50 methods in the class.

2. Extract the 6 FAR-only test methods into a new
   HelixClusterManagerFullAutoTest class. They previously gated on
   `assumeTrue(... && fullAutoCompatible)` and only ran on 1-2 rows out
   of 11 — wasting setup on the other 9. The new class is non-parameterized
   with a single FAR-compatible MockHelixCluster, so each test runs once
   instead of being skipped 9-10 times. Total FAR coverage is identical.

3. Revert maxParallelForks=4 + per-fork ZK port derivation. The CI run
   on PR #3241 with 4-way forks ran longer than the serial baseline,
   indicating CPU/memory contention on a 4-vCPU GitHub-hosted runner
   outweighed the parallelism win.

Local runs confirmed: parent class smoke test passes 9 parameterized
variants, new class passes all 6 FAR tests with zero skips.
